requirements.txt
networkx~=2.5
matplotlib~=3.3.4
numpy~=1.20.1
PyYAML~=5.4.1
pandas~=1.2.4
schema~=0.7.4